Test for config.data.iorfile before trying to unlink it
git-svn-id: svn+ssh://svn.gna.org/svn/advene/advene/trunk@646 3fb6f576-06d4-0310-ad11-a4b695aa9f05

+++ lib/advene/core/controller.py
@@ -115,7 +115,8 @@ class AdveneController:
             try:
                 # Kill spurious vlc player
                 os.system("/usr/bin/killall -9 vlc")
-                os.unlink(config.data.iorfile)
+                if os.access(config.data.iorfile, os.R_OK):
+                    os.unlink(config.data.iorfile)
             except OSError:
                 pass
